Return a single P matrix from a gt_admix object

This function retrieves a single P matrix from a gt_admix object based on the specified k value and run number.

Usage

get_p_matrix(x, ..., k, run)

Arguments

x

A gt_admix object containing P matrices

...

Not used

k

The k value of the desired P matrix

run

The run number of the desired P matrix

Value

A single P matrix from the gt_admix object

Examples

# Read example gt_admix object
admix_obj <-
  readRDS(system.file("extdata", "anolis", "anole_adm_k3.rds",
    package = "tidypopgen"
  ))

# Extract a P matrix
get_p_matrix(admix_obj, k = 3, run = 1)
#>            V1       V2       V3
#> 1    0.048612 0.999990 0.999990
#> 2    0.999990 0.837451 0.770223
#> 3    0.033211 0.999990 0.999990
#> 4    0.999990 0.000010 0.476634
#> 5    0.999990 0.561647 0.999990
#> 6    0.999990 0.999990 0.717089
#> 7    0.999990 0.105753 0.909817
#> 8    0.999990 0.999990 0.718884
#> 9    0.999990 0.125073 0.691520
#> 10   0.999990 0.335245 0.970162
#> 11   0.999990 0.174710 0.999990
#> 12   0.999990 0.999990 0.678217
#> 13   0.999990 0.801682 0.459852
#> 14   0.999990 0.416884 0.999990
#> 15   0.999990 0.000010 0.485328
#> 16   0.999990 0.000010 0.999990
#> 17   0.999990 0.000010 0.346382
#> 18   0.999990 0.000010 0.685250
#> 19   0.999990 0.639632 0.999990
#> 20   0.999990 0.514032 0.999990
#> 21   0.000010 0.999990 0.827757
#> 22   0.999990 0.131971 0.852027
#> 23   0.999990 0.999990 0.668089
#> 24   0.900621 0.000010 0.999990
#> 25   0.900505 0.000010 0.999990
#> 26   0.900621 0.000010 0.999990
#> 27   0.900621 0.000010 0.999990
#> 28   0.999990 0.068642 0.999990
#> 29   0.275223 0.999990 0.958552
#> 30   0.347394 0.999990 0.999990
#> 31   0.967079 0.000010 0.999990
#> 32   0.999990 0.999990 0.132778
#> 33   0.999990 0.000010 0.837993
#> 34   0.999990 0.328316 0.999990
#> 35   0.999990 0.375847 0.999990
#> 36   0.999990 0.000010 0.602393
#> 37   0.999990 0.999990 0.610375
#> 38   0.999990 0.288723 0.999990
#> 39   0.999990 0.000010 0.854501
#> 40   0.999990 0.335313 0.999990
#> 41   0.999990 0.999990 0.182406
#> 42   0.999990 0.000010 0.604867
#> 43   0.999990 0.000010 0.999990
#> 44   0.999990 0.000010 0.999990
#> 45   0.999990 0.596966 0.999990
#> 46   0.999990 0.000010 0.999990
#> 47   0.999990 0.000010 0.999990
#> 48   0.999990 0.596966 0.999990
#> 49   0.999990 0.000010 0.999990
#> 50   0.999990 0.049367 0.999990
#> 51   0.999990 0.999990 0.523934
#> 52   0.999990 0.344617 0.655521
#> 53   0.999990 0.746841 0.625154
#> 54   0.999990 0.000010 0.999990
#> 55   0.000010 0.999990 0.999990
#> 56   0.000010 0.999990 0.999990
#> 57   0.999990 0.000010 0.943597
#> 58   0.969131 0.386538 0.999990
#> 59   0.000010 0.999990 0.999990
#> 60   0.999990 0.000010 0.999990
#> 61   0.999990 0.419175 0.999990
#> 62   0.999990 0.076907 0.999990
#> 63   0.999990 0.162072 0.999990
#> 64   0.999990 0.999990 0.608016
#> 65   0.999990 0.169368 0.831542
#> 66   0.999990 0.165248 0.999990
#> 67   0.999990 0.000010 0.668063
#> 68   0.999990 0.515023 0.999990
#> 69   0.999990 0.000010 0.668063
#> 70   0.999990 0.000010 0.999990
#> 71   0.999990 0.071412 0.999990
#> 72   0.999990 0.999990 0.678757
#> 73   0.999990 0.000010 0.999990
#> 74   0.999990 0.999990 0.698978
#> 75   0.999990 0.469798 0.259533
#> 76   0.999990 0.375847 0.999990
#> 77   0.999990 0.000010 0.456997
#> 78   0.999990 0.144160 0.999990
#> 79   0.999990 0.535940 0.931303
#> 80   0.688474 0.515613 0.999990
#> 81   0.999990 0.000010 0.999990
#> 82   0.999990 0.380531 0.999990
#> 83   0.999990 0.256660 0.999990
#> 84   0.999990 0.099258 0.876406
#> 85   0.616057 0.999990 0.284705
#> 86   0.999990 0.294474 0.999990
#> 87   0.999990 0.469108 0.999990
#> 88   0.999990 0.260576 0.999990
#> 89   0.999990 0.521198 0.756456
#> 90   0.999990 0.000010 0.703064
#> 91   0.999990 0.000010 0.625649
#> 92   0.000010 0.999990 0.810944
#> 93   0.000010 0.999990 0.810944
#> 94   0.999990 0.295801 0.999990
#> 95   0.999990 0.050093 0.857528
#> 96   0.999990 0.679943 0.939650
#> 97   0.999990 0.561020 0.999990
#> 98   0.051939 0.999990 0.999990
#> 99   0.854262 0.957593 0.446715
#> 100  0.000010 0.999990 0.999990
#> 101  0.999990 0.000010 0.814765
#> 102  0.999990 0.559501 0.909287
#> 103  0.999990 0.721156 0.909239
#> 104  0.999990 0.514032 0.999990
#> 105  0.999990 0.183151 0.867017
#> 106  0.031782 0.999990 0.999990
#> 107  0.999990 0.074969 0.376982
#> 108  0.999990 0.162070 0.999990
#> 109  0.999990 0.999990 0.711043
#> 110  0.410293 0.957570 0.835815
#> 111  0.999990 0.000010 0.999990
#> 112  0.999990 0.000010 0.999990
#> 113  0.999990 0.000010 0.999990
#> 114  0.999990 0.466753 0.999990
#> 115  0.999990 0.415766 0.999990
#> 116  0.999990 0.999990 0.270395
#> 117  0.999990 0.515023 0.999990
#> 118  0.999990 0.959415 0.370706
#> 119  0.000010 0.999990 0.999990
#> 120  0.999990 0.302514 0.999990
#> 121  0.999990 0.517676 0.999990
#> 122  0.999990 0.000010 0.999990
#> 123  0.999990 0.000010 0.999990
#> 124  0.999990 0.560105 0.765791
#> 125  0.999990 0.000010 0.915067
#> 126  0.999990 0.344403 0.999990
#> 127  0.999990 0.344403 0.999990
#> 128  0.999990 0.038438 0.999990
#> 129  0.999990 0.000010 0.999990
#> 130  0.999990 0.000010 0.999990
#> 131  0.999990 0.000010 0.850175
#> 132  0.999990 0.182190 0.939709
#> 133  0.999990 0.953767 0.464839
#> 134  0.999990 0.605699 0.999990
#> 135  0.999990 0.000010 0.787129
#> 136  0.866669 0.999990 0.468969
#> 137  0.999990 0.000010 0.643417
#> 138  0.000010 0.999990 0.904410
#> 139  0.999990 0.000010 0.708804
#> 140  0.999990 0.000010 0.999990
#> 141  0.999990 0.440726 0.745239
#> 142  0.999990 0.999990 0.399495
#> 143  0.999990 0.999990 0.430931
#> 144  0.100547 0.999990 0.999990
#> 145  0.999990 0.917354 0.244776
#> 146  0.999990 0.000010 0.510772
#> 147  0.999990 0.957766 0.730184
#> 148  0.409860 0.999990 0.999990
#> 149  0.999990 0.000010 0.396582
#> 150  0.000010 0.999990 0.999990
#> 151  0.999990 0.999990 0.570089
#> 152  0.652832 0.999990 0.999990
#> 153  0.999990 0.550824 0.999990
#> 154  0.999990 0.843455 0.709648
#> 155  0.999990 0.843455 0.709648
#> 156  0.166674 0.999990 0.514208
#> 157  0.167157 0.999990 0.486349
#> 158  0.167179 0.999990 0.485315
#> 159  0.999990 0.208222 0.850440
#> 160  0.999990 0.999990 0.658225
#> 161  0.861593 0.694844 0.294181
#> 162  0.999990 0.000010 0.808827
#> 163  0.999990 0.000010 0.999990
#> 164  0.000010 0.999990 0.955633
#> 165  0.000010 0.999990 0.971158
#> 166  0.999990 0.567760 0.999990
#> 167  0.999990 0.417002 0.999990
#> 168  0.000010 0.827580 0.944318
#> 169  0.999990 0.558151 0.854535
#> 170  0.091985 0.999990 0.831712
#> 171  0.849785 0.999990 0.469028
#> 172  0.999990 0.999990 0.526063
#> 173  0.999990 0.000010 0.999990
#> 174  0.037335 0.999990 0.868742
#> 175  0.037276 0.999990 0.876849
#> 176  0.999990 0.000010 0.999990
#> 177  0.999990 0.105940 0.999990
#> 178  0.999990 0.105940 0.999990
#> 179  0.999990 0.000010 0.999990
#> 180  0.999990 0.999990 0.000010
#> 181  0.999990 0.000010 0.999990
#> 182  0.999990 0.000010 0.999990
#> 183  0.999990 0.000010 0.999990
#> 184  0.999990 0.999990 0.729027
#> 185  0.999990 0.038218 0.868367
#> 186  0.999990 0.472302 0.999990
#> 187  0.000010 0.999990 0.999990
#> 188  0.072803 0.999990 0.925630
#> 189  0.999990 0.000010 0.999990
#> 190  0.999990 0.000010 0.907002
#> 191  0.999990 0.500142 0.817751
#> 192  0.000010 0.999990 0.778672
#> 193  0.000010 0.999990 0.999990
#> 194  0.000010 0.999990 0.999990
#> 195  0.000010 0.999990 0.999990
#> 196  0.000010 0.999990 0.999990
#> 197  0.999990 0.516536 0.999990
#> 198  0.000010 0.999990 0.816185
#> 199  0.000010 0.873207 0.853425
#> 200  0.999990 0.279973 0.999990
#> 201  0.999990 0.999990 0.634823
#> 202  0.999990 0.000010 0.999990
#> 203  0.999990 0.102728 0.999990
#> 204  0.999990 0.602253 0.999990
#> 205  0.000010 0.999990 0.969391
#> 206  0.999990 0.957700 0.751467
#> 207  0.000010 0.999990 0.969391
#> 208  0.999990 0.279975 0.999990
#> 209  0.956135 0.999990 0.560701
#> 210  0.968495 0.515021 0.999990
#> 211  0.999990 0.328318 0.999990
#> 212  0.999990 0.000010 0.999990
#> 213  0.999990 0.145199 0.999990
#> 214  0.999990 0.000010 0.999990
#> 215  0.999990 0.000010 0.999990
#> 216  0.999990 0.000010 0.999990
#> 217  0.999990 0.576409 0.999990
#> 218  0.999990 0.000010 0.473852
#> 219  0.965902 0.999990 0.421242
#> 220  0.968990 0.210476 0.999990
#> 221  0.999990 0.210478 0.999990
#> 222  0.999990 0.210478 0.999990
#> 223  0.999990 0.000010 0.999990
#> 224  0.999990 0.577048 0.969313
#> 225  0.999990 0.183827 0.940519
#> 226  0.999990 0.000010 0.999990
#> 227  0.999990 0.000010 0.377578
#> 228  0.999990 0.129412 0.999990
#> 229  0.999990 0.000010 0.999990
#> 230  0.725502 0.999990 0.125618
#> 231  0.725502 0.999990 0.125618
#> 232  0.000010 0.741713 0.851621
#> 233  0.999990 0.999990 0.177166
#> 234  0.999990 0.562745 0.999990
#> 235  0.999990 0.513764 0.999990
#> 236  0.999990 0.000010 0.753858
#> 237  0.999990 0.000010 0.413284
#> 238  0.999990 0.861305 0.680848
#> 239  0.999990 0.174711 0.999990
#> 240  0.999990 0.000010 0.409729
#> 241  0.999990 0.000010 0.999990
#> 242  0.999990 0.000010 0.999990
#> 243  0.999990 0.315301 0.999990
#> 244  0.999990 0.585722 0.999990
#> 245  0.999990 0.633506 0.999990
#> 246  0.999990 0.633506 0.999990
#> 247  0.999990 0.220862 0.999990
#> 248  0.999990 0.220862 0.999990
#> 249  0.999990 0.000010 0.694164
#> 250  0.999990 0.999990 0.613571
#> 251  0.999990 0.000010 0.709210
#> 252  0.999990 0.000010 0.709210
#> 253  0.662827 0.714058 0.999990
#> 254  0.999990 0.147170 0.942996
#> 255  0.999990 0.878999 0.696670
#> 256  0.682662 0.881307 0.958581
#> 257  0.000010 0.914677 0.846852
#> 258  0.999990 0.999990 0.644819
#> 259  0.999990 0.000010 0.665301
#> 260  0.999990 0.151795 0.307907
#> 261  0.999990 0.999990 0.701810
#> 262  0.999990 0.561543 0.971157
#> 263  0.000010 0.935048 0.754686
#> 264  0.999990 0.999990 0.153325
#> 265  0.944716 0.999990 0.160456
#> 266  0.952943 0.234144 0.866339
#> 267  0.999990 0.000010 0.964612
#> 268  0.999990 0.000010 0.674912
#> 269  0.000010 0.999990 0.972211
#> 270  0.999990 0.375848 0.999990
#> 271  0.999990 0.119746 0.659259
#> 272  0.999990 0.269366 0.434778
#> 273  0.999990 0.344402 0.999990
#> 274  0.999990 0.285364 0.999990
#> 275  0.999990 0.878541 0.999990
#> 276  0.999990 0.448037 0.961685
#> 277  0.887529 0.000010 0.824114
#> 278  0.999990 0.999990 0.448567
#> 279  0.999990 0.328317 0.999990
#> 280  0.999990 0.040652 0.616039
#> 281  0.000010 0.999990 0.700512
#> 282  0.652236 0.999990 0.999990
#> 283  0.999990 0.999990 0.594896
#> 284  0.999990 0.000010 0.544303
#> 285  0.999990 0.000010 0.927907
#> 286  0.999990 0.000010 0.879535
#> 287  0.999990 0.999990 0.694498
#> 288  0.999990 0.000010 0.415671
#> 289  0.999990 0.999990 0.412560
#> 290  0.999990 0.044949 0.999990
#> 291  0.999990 0.000010 0.999990
#> 292  0.999796 0.688972 0.810406
#> 293  0.999990 0.250358 0.999990
#> 294  0.154877 0.999990 0.810580
#> 295  0.845316 0.999990 0.179597
#> 296  0.999990 0.000010 0.999990
#> 297  0.610618 0.999990 0.410662
#> 298  0.999990 0.999990 0.405206
#> 299  0.999990 0.999990 0.292376
#> 300  0.999990 0.158372 0.999990
#> 301  0.999990 0.158372 0.999990
#> 302  0.000010 0.999990 0.652525
#> 303  0.999990 0.000010 0.999990
#> 304  0.999990 0.000010 0.999990
#> 305  0.999990 0.480747 0.825883
#> 306  0.999990 0.000010 0.999990
#> 307  0.999990 0.999990 0.000010
#> 308  0.000010 0.999990 0.873057
#> 309  0.999990 0.000010 0.999990
#> 310  0.999990 0.000010 0.409204
#> 311  0.948547 0.054950 0.962441
#> 312  0.948547 0.054950 0.962441
#> 313  0.999990 0.959717 0.577117
#> 314  0.999990 0.879161 0.579619
#> 315  0.999990 0.000010 0.966023
#> 316  0.999990 0.000010 0.999990
#> 317  0.999990 0.000010 0.674103
#> 318  0.000010 0.999990 0.999990
#> 319  0.999990 0.000010 0.999990
#> 320  0.999990 0.000010 0.999990
#> 321  0.247475 0.999990 0.717168
#> 322  0.885765 0.265639 0.694652
#> 323  0.000010 0.999990 0.999990
#> 324  0.999990 0.000010 0.838312
#> 325  0.999990 0.519435 0.999990
#> 326  0.999990 0.520055 0.999990
#> 327  0.999990 0.688110 0.813214
#> 328  0.999990 0.300847 0.999990
#> 329  0.999990 0.000010 0.341661
#> 330  0.810455 0.999990 0.173903
#> 331  0.182748 0.961137 0.923945
#> 332  0.999990 0.281639 0.999990
#> 333  0.999990 0.000010 0.866871
#> 334  0.999990 0.000010 0.926996
#> 335  0.999990 0.999990 0.596592
#> 336  0.875677 0.000010 0.999990
#> 337  0.999990 0.597343 0.999990
#> 338  0.999990 0.000010 0.603204
#> 339  0.836057 0.999990 0.000010
#> 340  0.999990 0.541416 0.694746
#> 341  0.999990 0.555955 0.999990
#> 342  0.153924 0.999990 0.999990
#> 343  0.000010 0.999990 0.794270
#> 344  0.999990 0.165249 0.999990
#> 345  0.999990 0.000010 0.999990
#> 346  0.999990 0.000010 0.999990
#> 347  0.999990 0.999990 0.691514
#> 348  0.999990 0.907418 0.032656
#> 349  0.000010 0.999990 0.965916
#> 350  0.999990 0.661675 0.908887
#> 351  0.999990 0.661675 0.908887
#> 352  0.999990 0.621166 0.122805
#> 353  0.999990 0.000010 0.999990
#> 354  0.999990 0.079614 0.999990
#> 355  0.999990 0.079614 0.999990
#> 356  0.366615 0.279997 0.999990
#> 357  0.999990 0.514299 0.999990
#> 358  0.999990 0.999990 0.476416
#> 359  0.999990 0.282245 0.939952
#> 360  0.999990 0.999990 0.717036
#> 361  0.999990 0.375848 0.999990
#> 362  0.999990 0.321805 0.788220
#> 363  0.999990 0.469526 0.999990
#> 364  0.999990 0.999990 0.645347
#> 365  0.999990 0.555504 0.999990
#> 366  0.999990 0.557570 0.999990
#> 367  0.999990 0.999990 0.698978
#> 368  0.999990 0.000010 0.999990
#> 369  0.000010 0.960524 0.978221
#> 370  0.057848 0.999990 0.751192
#> 371  0.069728 0.999990 0.774352
#> 372  0.999990 0.313320 0.912608
#> 373  0.999990 0.087477 0.814344
#> 374  0.977132 0.000010 0.970631
#> 375  0.123159 0.999990 0.971702
#> 376  0.999990 0.469109 0.999990
#> 377  0.999990 0.287169 0.999990
#> 378  0.999990 0.000010 0.933508
#> 379  0.999990 0.513764 0.999990
#> 380  0.000010 0.999990 0.999990
#> 381  0.999990 0.283958 0.295232
#> 382  0.991697 0.039638 0.999065
#> 383  0.999990 0.815501 0.834539
#> 384  0.999990 0.815501 0.834539
#> 385  0.999990 0.000010 0.838074
#> 386  0.999990 0.182360 0.999990
#> 387  0.999990 0.000010 0.862814
#> 388  0.999990 0.000010 0.686923
#> 389  0.712052 0.999990 0.999990
#> 390  0.999990 0.000010 0.999990
#> 391  0.999990 0.422742 0.999990
#> 392  0.999990 0.000010 0.999990
#> 393  0.000010 0.999990 0.826755
#> 394  0.999990 0.649099 0.224162
#> 395  0.999990 0.999990 0.641208
#> 396  0.598257 0.999990 0.929189
#> 397  0.999990 0.000010 0.999990
#> 398  0.000010 0.999990 0.999990
#> 399  0.999990 0.000010 0.999990
#> 400  0.000010 0.951996 0.941433
#> 401  0.999990 0.246176 0.802543
#> 402  0.999990 0.655718 0.860045
#> 403  0.999990 0.999990 0.586661
#> 404  0.487520 0.263501 0.561569
#> 405  0.999990 0.597102 0.550741
#> 406  0.191081 0.999990 0.939236
#> 407  0.892302 0.999990 0.466366
#> 408  0.646959 0.999990 0.897852
#> 409  0.999990 0.261309 0.968741
#> 410  0.999990 0.213371 0.937262
#> 411  0.975323 0.416781 0.412016
#> 412  0.999990 0.000010 0.771590
#> 413  0.999990 0.088216 0.876248
#> 414  0.999990 0.281639 0.999990
#> 415  0.999990 0.999990 0.443691
#> 416  0.999990 0.999990 0.471669
#> 417  0.999990 0.000010 0.914737
#> 418  0.999990 0.999990 0.443691
#> 419  0.751706 0.945848 0.922072
#> 420  0.999990 0.422742 0.999990
#> 421  0.999990 0.000010 0.590259
#> 422  0.999990 0.647140 0.609412
#> 423  0.999990 0.357000 0.968023
#> 424  0.460303 0.999990 0.914788
#> 425  0.999990 0.201551 0.999990
#> 426  0.999990 0.473173 0.999990
#> 427  0.999990 0.000010 0.891198
#> 428  0.999990 0.117980 0.999990
#> 429  0.999990 0.079614 0.999990
#> 430  0.999990 0.000010 0.999990
#> 431  0.999990 0.518272 0.969680
#> 432  0.999990 0.000010 0.763431
#> 433  0.999990 0.000010 0.782583
#> 434  0.999990 0.000010 0.553887
#> 435  0.000010 0.999990 0.999990
#> 436  0.999990 0.000010 0.999990
#> 437  0.000010 0.999990 0.796848
#> 438  0.579842 0.999990 0.999990
#> 439  0.999990 0.381149 0.407582
#> 440  0.999990 0.288723 0.999990
#> 441  0.999990 0.166535 0.943568
#> 442  0.000010 0.999990 0.724145
#> 443  0.029540 0.999990 0.999990
#> 444  0.968335 0.999990 0.634422
#> 445  0.999990 0.414790 0.999990
#> 446  0.999990 0.000010 0.999990
#> 447  0.727441 0.519826 0.335840
#> 448  0.000010 0.999990 0.999990
#> 449  0.788447 0.596970 0.999990
#> 450  0.999990 0.000010 0.999990
#> 451  0.999990 0.000010 0.999990
#> 452  0.000010 0.999990 0.936424
#> 453  0.999990 0.000010 0.999990
#> 454  0.999990 0.000010 0.999990
#> 455  0.999990 0.000010 0.999990
#> 456  0.999990 0.044949 0.999990
#> 457  0.999990 0.999990 0.697908
#> 458  0.999990 0.000010 0.999990
#> 459  0.999990 0.213481 0.999990
#> 460  0.000010 0.999990 0.960352
#> 461  0.999990 0.358507 0.800680
#> 462  0.000010 0.999990 0.732984
#> 463  0.999990 0.507362 0.999990
#> 464  0.999990 0.787326 0.310508
#> 465  0.999990 0.742737 0.323214
#> 466  0.000010 0.999990 0.953046
#> 467  0.999990 0.177426 0.999990
#> 468  0.999990 0.999990 0.138253
#> 469  0.999990 0.000010 0.859693
#> 470  0.999990 0.472301 0.999990
#> 471  0.999990 0.999990 0.693390
#> 472  0.999990 0.000010 0.807384
#> 473  0.999990 0.623586 0.935488
#> 474  0.999990 0.713723 0.824993
#> 475  0.555065 0.999990 0.819041
#> 476  0.999990 0.999990 0.262370
#> 477  0.999990 0.000010 0.769680
#> 478  0.999990 0.258107 0.999990
#> 479  0.999990 0.000010 0.445722
#> 480  0.999990 0.195235 0.937181
#> 481  0.999990 0.194352 0.999990
#> 482  0.000010 0.999990 0.952696
#> 483  0.000010 0.999990 0.952696
#> 484  0.087721 0.999990 0.999990
#> 485  0.999990 0.386539 0.999990
#> 486  0.999990 0.375848 0.999990
#> 487  0.999990 0.469109 0.999990
#> 488  0.999990 0.000010 0.841984
#> 489  0.999990 0.000010 0.999990
#> 490  0.999990 0.000010 0.999990
#> 491  0.119121 0.999990 0.659241
#> 492  0.999990 0.502329 0.999990
#> 493  0.999990 0.999990 0.578356
#> 494  0.385838 0.999990 0.999990
#> 495  0.889423 0.843113 0.478583
#> 496  0.999990 0.115738 0.602063
#> 497  0.999990 0.000010 0.599949
#> 498  0.000010 0.999990 0.865717
#> 499  0.899600 0.999990 0.731773
#> 500  0.999990 0.520055 0.999990
#> 501  0.999990 0.999990 0.399441
#> 502  0.999990 0.514299 0.999990
#> 503  0.000010 0.999990 0.999990
#> 504  0.999990 0.259060 0.999990
#> 505  0.000010 0.999990 0.811886
#> 506  0.999990 0.471717 0.999990
#> 507  0.999990 0.060255 0.999990
#> 508  0.999990 0.227241 0.698126
#> 509  0.999990 0.335314 0.999990
#> 510  0.999990 0.335314 0.999990
#> 511  0.999990 0.033315 0.999990
#> 512  0.999990 0.174258 0.999990
#> 513  0.114414 0.999990 0.549711
#> 514  0.114414 0.999990 0.549711
#> 515  0.114414 0.999990 0.549711
#> 516  0.886434 0.559696 0.438981
#> 517  0.178655 0.999990 0.937037
#> 518  0.999990 0.999990 0.691370
#> 519  0.999990 0.000010 0.999990
#> 520  0.860571 0.056690 0.733846
#> 521  0.999990 0.049368 0.999990
#> 522  0.999990 0.559963 0.378026
#> 523  0.999990 0.590172 0.976284
#> 524  0.999990 0.000010 0.724327
#> 525  0.999990 0.601164 0.999990
#> 526  0.999990 0.095269 0.375615
#> 527  0.999990 0.000010 0.752206
#> 528  0.999990 0.518621 0.999990
#> 529  0.999990 0.000010 0.414847
#> 530  0.000010 0.999990 0.855769
#> 531  0.999990 0.515361 0.913370
#> 532  0.999990 0.089490 0.999990
#> 533  0.999990 0.000010 0.921949
#> 534  0.999990 0.000010 0.999990
#> 535  0.999990 0.000010 0.999990
#> 536  0.999990 0.999990 0.228634
#> 537  0.295210 0.999990 0.738029
#> 538  0.999990 0.420116 0.999990
#> 539  0.999990 0.172817 0.951508
#> 540  0.999990 0.196688 0.999990
#> 541  0.999990 0.232708 0.999990
#> 542  0.999990 0.232708 0.999990
#> 543  0.902594 0.999990 0.311719
#> 544  0.000010 0.999990 0.968091
#> 545  0.999990 0.275617 0.339990
#> 546  0.999990 0.268341 0.441702
#> 547  0.999990 0.999990 0.541090
#> 548  0.999990 0.999990 0.548198
#> 549  0.999990 0.349077 0.797363
#> 550  0.999990 0.264489 0.999990
#> 551  0.999990 0.395606 0.657427
#> 552  0.999990 0.961044 0.737214
#> 553  0.999990 0.000010 0.597645
#> 554  0.999990 0.390638 0.999990
#> 555  0.861138 0.958781 0.479027
#> 556  0.137427 0.999990 0.572835
#> 557  0.985619 0.000010 0.948692
#> 558  0.985619 0.000010 0.948692
#> 559  0.260006 0.999990 0.667368
#> 560  0.999990 0.176081 0.999990
#> 561  0.999990 0.557570 0.999990
#> 562  0.999990 0.000010 0.999990
#> 563  0.999990 0.000010 0.602398
#> 564  0.999990 0.999990 0.610375
#> 565  0.999990 0.000010 0.633956
#> 566  0.000010 0.741475 0.918487
#> 567  0.999990 0.399011 0.415735
#> 568  0.038835 0.999990 0.698576
#> 569  0.999990 0.999990 0.146449
#> 570  0.999990 0.200047 0.999990
#> 571  0.999990 0.200047 0.999990
#> 572  0.999990 0.330969 0.939734
#> 573  0.999990 0.330969 0.939734
#> 574  0.452802 0.999990 0.999990
#> 575  0.000010 0.999990 0.999990
#> 576  0.999990 0.000010 0.968158
#> 577  0.999990 0.000010 0.999990
#> 578  0.999990 0.000010 0.999990
#> 579  0.999990 0.472035 0.999990
#> 580  0.999990 0.000010 0.905487
#> 581  0.999990 0.000010 0.445702
#> 582  0.999990 0.134319 0.932396
#> 583  0.999990 0.000010 0.999990
#> 584  0.999990 0.000010 0.999990
#> 585  0.999990 0.999990 0.654306
#> 586  0.999990 0.000010 0.528582
#> 587  0.999990 0.999990 0.633684
#> 588  0.999990 0.000010 0.655499
#> 589  0.193870 0.999990 0.936862
#> 590  0.000010 0.999990 0.936952
#> 591  0.999990 0.999990 0.568423
#> 592  0.999990 0.051209 0.999990
#> 593  0.999990 0.000010 0.999990
#> 594  0.999990 0.000010 0.999990
#> 595  0.582099 0.999990 0.000010
#> 596  0.999990 0.999990 0.698957
#> 597  0.065912 0.999990 0.999990
#> 598  0.000010 0.999990 0.999990
#> 599  0.999990 0.000010 0.431619
#> 600  0.999990 0.279974 0.999990
#> 601  0.999990 0.000010 0.923479
#> 602  0.000010 0.999990 0.999990
#> 603  0.240705 0.999990 0.601846
#> 604  0.999990 0.652091 0.902057
#> 605  0.966566 0.999990 0.438907
#> 606  0.999990 0.000010 0.999990
#> 607  0.999990 0.000010 0.502637
#> 608  0.999990 0.000010 0.999990
#> 609  0.999990 0.120686 0.999990
#> 610  0.598915 0.999990 0.670033
#> 611  0.999990 0.049367 0.999990
#> 612  0.999990 0.000010 0.740541
#> 613  0.680297 0.488819 0.219818
#> 614  0.999990 0.165249 0.999990
#> 615  0.837463 0.999990 0.301122
#> 616  0.999990 0.426595 0.999990
#> 617  0.999990 0.000010 0.873137
#> 618  0.999990 0.999990 0.524765
#> 619  0.999990 0.000010 0.718375
#> 620  0.310989 0.999990 0.999990
#> 621  0.999990 0.558012 0.999990
#> 622  0.999990 0.000010 0.999990
#> 623  0.999990 0.000010 0.999990
#> 624  0.999990 0.999990 0.311926
#> 625  0.000010 0.999990 0.999990
#> 626  0.999990 0.999990 0.322580
#> 627  0.999990 0.999990 0.311926
#> 628  0.000010 0.999990 0.999990
#> 629  0.999990 0.452485 0.999990
#> 630  0.999990 0.570306 0.999990
#> 631  0.999990 0.462116 0.999990
#> 632  0.999990 0.000010 0.999990
#> 633  0.999990 0.279974 0.999990
#> 634  0.999990 0.000010 0.800774
#> 635  0.999990 0.068642 0.999990
#> 636  0.999990 0.605698 0.999990
#> 637  0.000010 0.999990 0.841928
#> 638  0.000010 0.999990 0.999990
#> 639  0.999990 0.101580 0.999990
#> 640  0.999990 0.000010 0.999990
#> 641  0.999990 0.124497 0.999990
#> 642  0.999990 0.040558 0.557351
#> 643  0.999990 0.081546 0.999990
#> 644  0.000010 0.999990 0.828407
#> 645  0.680000 0.999990 0.999990
#> 646  0.000010 0.999990 0.775561
#> 647  0.999990 0.074477 0.851361
#> 648  0.999990 0.000010 0.389508
#> 649  0.174261 0.999990 0.999990
#> 650  0.298924 0.999990 0.999990
#> 651  0.130451 0.999990 0.951567
#> 652  0.999990 0.556541 0.999990
#> 653  0.000010 0.999990 0.771753
#> 654  0.999990 0.287168 0.999990
#> 655  0.999990 0.254922 0.999990
#> 656  0.999990 0.000010 0.502437
#> 657  0.965491 0.999990 0.075345
#> 658  0.034531 0.957265 0.920834
#> 659  0.965477 0.616900 0.077423
#> 660  0.034532 0.953394 0.920660
#> 661  0.965498 0.836493 0.097809
#> 662  0.999990 0.897483 0.555872
#> 663  0.999990 0.316957 0.999990
#> 664  0.999990 0.000010 0.864098
#> 665  0.999990 0.508960 0.933105
#> 666  0.999990 0.569441 0.999990
#> 667  0.999990 0.108751 0.999990
#> 668  0.999990 0.000010 0.837992
#> 669  0.999990 0.516735 0.999990
#> 670  0.999990 0.000010 0.868458
#> 671  0.999990 0.000010 0.814015
#> 672  0.999990 0.182360 0.999990
#> 673  0.999990 0.182360 0.999990
#> 674  0.000010 0.999990 0.999990
#> 675  0.000010 0.812384 0.929035
#> 676  0.999990 0.000010 0.901693
#> 677  0.999990 0.129148 0.862095
#> 678  0.999990 0.511361 0.999990
#> 679  0.884065 0.000010 0.830740
#> 680  0.228851 0.999990 0.999990
#> 681  0.999990 0.000010 0.940219
#> 682  0.999990 0.186693 0.238681
#> 683  0.999990 0.999990 0.432893
#> 684  0.999990 0.000010 0.533048
#> 685  0.999990 0.000010 0.570206
#> 686  0.642047 0.999990 0.944120
#> 687  0.999990 0.430086 0.999990
#> 688  0.999990 0.560540 0.999990
#> 689  0.999990 0.000010 0.526796
#> 690  0.999990 0.000010 0.528814
#> 691  0.000010 0.999990 0.639988
#> 692  0.999990 0.842439 0.812347
#> 693  0.000010 0.999990 0.965916
#> 694  0.999990 0.146514 0.999990
#> 695  0.999990 0.000010 0.509688
#> 696  0.999990 0.999990 0.638529
#> 697  0.999990 0.000010 0.999990
#> 698  0.903333 0.000010 0.570478
#> 699  0.999990 0.357250 0.989143
#> 700  0.999990 0.638173 0.999990
#> 701  0.000010 0.999990 0.655692
#> 702  0.999990 0.384283 0.999990
#> 703  0.999990 0.999990 0.637404
#> 704  0.668675 0.999990 0.524197
#> 705  0.999990 0.515604 0.999990
#> 706  0.999990 0.000010 0.939553
#> 707  0.000010 0.999990 0.933859
#> 708  0.999990 0.000010 0.798655
#> 709  0.554195 0.999990 0.999990
#> 710  0.999990 0.000010 0.886850
#> 711  0.999990 0.999990 0.557583
#> 712  0.999990 0.471062 0.908484
#> 713  0.999990 0.234226 0.936113
#> 714  0.999990 0.234226 0.936113
#> 715  0.999990 0.091539 0.634259
#> 716  0.931285 0.610710 0.999979
#> 717  0.999990 0.000010 0.999990
#> 718  0.999990 0.083315 0.999990
#> 719  0.999990 0.083315 0.999990
#> 720  0.999990 0.999990 0.630011
#> 721  0.999990 0.511304 0.999990
#> 722  0.999990 0.999990 0.175518
#> 723  0.999990 0.999990 0.000010
#> 724  0.999990 0.431066 0.914169
#> 725  0.999990 0.077186 0.999990
#> 726  0.245868 0.999984 0.548946
#> 727  0.999990 0.423597 0.999990
#> 728  0.999990 0.697928 0.214931
#> 729  0.999990 0.272147 0.310589
#> 730  0.999990 0.560540 0.999990
#> 731  0.999990 0.999990 0.451873
#> 732  0.999990 0.429325 0.999990
#> 733  0.999990 0.000010 0.883599
#> 734  0.999990 0.294912 0.681942
#> 735  0.443896 0.999990 0.643455
#> 736  0.999990 0.555955 0.999990
#> 737  0.999990 0.999990 0.576382
#> 738  0.999990 0.000010 0.517681
#> 739  0.999990 0.466754 0.999990
#> 740  0.999990 0.999990 0.700196
#> 741  0.000010 0.999990 0.999990
#> 742  0.999990 0.999990 0.700196
#> 743  0.888531 0.683340 0.469753
#> 744  0.878911 0.773991 0.786426
#> 745  0.999990 0.000010 0.999990
#> 746  0.491091 0.999990 0.209142
#> 747  0.038950 0.955836 0.623186
#> 748  0.999990 0.999990 0.022177
#> 749  0.999990 0.999990 0.022177
#> 750  0.999990 0.129809 0.999990
#> 751  0.999990 0.999990 0.711434
#> 752  0.999990 0.999990 0.711434
#> 753  0.999990 0.212591 0.905506
#> 754  0.999990 0.165249 0.999990
#> 755  0.206006 0.999990 0.999990
#> 756  0.999990 0.457270 0.999990
#> 757  0.999990 0.041542 0.839000
#> 758  0.999990 0.000010 0.999990
#> 759  0.999990 0.000010 0.999990
#> 760  0.999990 0.000010 0.999990
#> 761  0.999990 0.044944 0.999990
#> 762  0.999990 0.000010 0.999990
#> 763  0.999990 0.000010 0.999990
#> 764  0.000010 0.999990 0.787580
#> 765  0.000010 0.999990 0.787580
#> 766  0.999990 0.396491 0.325784
#> 767  0.999990 0.999990 0.297985
#> 768  0.999990 0.999990 0.297985
#> 769  0.999990 0.617456 0.627385
#> 770  0.999990 0.423427 0.967942
#> 771  0.999990 0.145198 0.999990
#> 772  0.999990 0.404990 0.999990
#> 773  0.999990 0.917716 0.110995
#> 774  0.208522 0.913966 0.999990
#> 775  0.999990 0.420117 0.999990
#> 776  0.999990 0.999990 0.432280
#> 777  0.000010 0.999990 0.626910
#> 778  0.999990 0.665241 0.569919
#> 779  0.999990 0.857066 0.259653
#> 780  0.999990 0.182361 0.999990
#> 781  0.761611 0.999990 0.861278
#> 782  0.999990 0.000010 0.999990
#> 783  0.999990 0.131483 0.967713
#> 784  0.999990 0.605698 0.999990
#> 785  0.999990 0.044944 0.999990
#> 786  0.999990 0.000010 0.999990
#> 787  0.000010 0.999990 0.816942
#> 788  0.999990 0.472301 0.999990
#> 789  0.999990 0.517677 0.999990
#> 790  0.679938 0.999990 0.999990
#> 791  0.999990 0.999990 0.277696
#> 792  0.999990 0.254922 0.999990
#> 793  0.999990 0.000010 0.999990
#> 794  0.999990 0.608659 0.942182
#> 795  0.999990 0.524818 0.433357
#> 796  0.999990 0.664605 0.442732
#> 797  0.999990 0.343290 0.999990
#> 798  0.999990 0.112788 0.509518
#> 799  0.999990 0.050368 0.999990
#> 800  0.999990 0.999990 0.439474
#> 801  0.999990 0.085806 0.999990
#> 802  0.999990 0.098973 0.911287
#> 803  0.999990 0.999990 0.689192
#> 804  0.999990 0.613197 0.848267
#> 805  0.999990 0.000010 0.663959
#> 806  0.999990 0.243215 0.910543
#> 807  0.999990 0.605699 0.999990
#> 808  0.999990 0.000010 0.999990
#> 809  0.999990 0.000010 0.999990
#> 810  0.999990 0.044949 0.999990
#> 811  0.999990 0.302514 0.999990
#> 812  0.068826 0.826862 0.851885
#> 813  0.903321 0.999990 0.165210
#> 814  0.000010 0.999990 0.999990
#> 815  0.000010 0.999990 0.999990
#> 816  0.999990 0.239826 0.999990
#> 817  0.999990 0.297630 0.999990
#> 818  0.999990 0.000010 0.373777
#> 819  0.999990 0.000010 0.402876
#> 820  0.616842 0.999990 0.077025
#> 821  0.999990 0.595728 0.999990
#> 822  0.999990 0.568937 0.999990
#> 823  0.999990 0.000010 0.999990
#> 824  0.999990 0.473921 0.177862
#> 825  0.000010 0.999990 0.967394
#> 826  0.999990 0.567761 0.999990
#> 827  0.000010 0.999990 0.967394
#> 828  0.999990 0.000010 0.999990
#> 829  0.999990 0.666742 0.897124
#> 830  0.032033 0.999990 0.999990
#> 831  0.999990 0.000010 0.999990
#> 832  0.999990 0.299474 0.999990
#> 833  0.999990 0.000010 0.999990
#> 834  0.590868 0.999990 0.999990
#> 835  0.999990 0.000010 0.999990
#> 836  0.999990 0.000010 0.943533
#> 837  0.999990 0.000010 0.943533
#> 838  0.999990 0.000010 0.999990
#> 839  0.999990 0.299474 0.999990
#> 840  0.324643 0.999990 0.999990
#> 841  0.728518 0.999990 0.514995
#> 842  0.999990 0.999990 0.712624
#> 843  0.999990 0.132352 0.830072
#> 844  0.999990 0.999990 0.706240
#> 845  0.999990 0.000010 0.999990
#> 846  0.999990 0.209814 0.999990
#> 847  0.999990 0.000010 0.999990
#> 848  0.999990 0.000010 0.999990
#> 849  0.860819 0.175059 0.844178
#> 850  0.885410 0.176697 0.652843
#> 851  0.999990 0.872878 0.999990
#> 852  0.999990 0.872878 0.999990
#> 853  0.999990 0.872878 0.999990
#> 854  0.999990 0.475234 0.999990
#> 855  0.999990 0.000010 0.999990
#> 856  0.999990 0.000010 0.999990
#> 857  0.999990 0.068642 0.999990
#> 858  0.999990 0.068642 0.999990
#> 859  0.999990 0.000010 0.999990
#> 860  0.000010 0.999990 0.623339
#> 861  0.000010 0.686411 0.642386
#> 862  0.000010 0.999990 0.623339
#> 863  0.520750 0.999990 0.999990
#> 864  0.999990 0.038438 0.999990
#> 865  0.999990 0.000010 0.999990
#> 866  0.999990 0.349283 0.999990
#> 867  0.999990 0.037326 0.924284
#> 868  0.999990 0.184536 0.913267
#> 869  0.999990 0.000010 0.782102
#> 870  0.274749 0.999990 0.972613
#> 871  0.999990 0.000010 0.999990
#> 872  0.000010 0.870647 0.999990
#> 873  0.999990 0.999990 0.000010
#> 874  0.999990 0.999990 0.000010
#> 875  0.999990 0.999990 0.000010
#> 876  0.000010 0.999990 0.802492
#> 877  0.999990 0.002448 0.593284
#> 878  0.999990 0.757757 0.549636
#> 879  0.999990 0.000010 0.559251
#> 880  0.999990 0.049368 0.999990
#> 881  0.999990 0.827790 0.691532
#> 882  0.999990 0.999990 0.667636
#> 883  0.999990 0.000010 0.741439
#> 884  0.161429 0.999990 0.999990
#> 885  0.999990 0.564255 0.089244
#> 886  0.999990 0.999990 0.392881
#> 887  0.999990 0.000010 0.937640
#> 888  0.999990 0.549287 0.999990
#> 889  0.777692 0.999990 0.764499
#> 890  0.999990 0.000010 0.999990
#> 891  0.999990 0.000010 0.999990
#> 892  0.999990 0.999990 0.640196
#> 893  0.999990 0.999990 0.640196
#> 894  0.999990 0.999990 0.531017
#> 895  0.999990 0.999990 0.164858
#> 896  0.999990 0.000010 0.410434
#> 897  0.999990 0.000010 0.410434
#> 898  0.999990 0.000010 0.999990
#> 899  0.999990 0.000010 0.834326
#> 900  0.999990 0.999990 0.273975
#> 901  0.000010 0.924028 0.833459
#> 902  0.999990 0.194738 0.971479
#> 903  0.999990 0.999990 0.732057
#> 904  0.335543 0.999990 0.999990
#> 905  0.999990 0.000010 0.999990
#> 906  0.999990 0.999990 0.041605
#> 907  0.999990 0.318440 0.999990
#> 908  0.999990 0.391521 0.999990
#> 909  0.999990 0.132929 0.504888
#> 910  0.999990 0.132929 0.504888
#> 911  0.999990 0.496665 0.797831
#> 912  0.999990 0.000010 0.721627
#> 913  0.957704 0.275625 0.786608
#> 914  0.954318 0.011197 0.852582
#> 915  0.999990 0.000010 0.999990
#> 916  0.309075 0.999990 0.513512
#> 917  0.000010 0.999990 0.883822
#> 918  0.000010 0.999990 0.677819
#> 919  0.999990 0.955116 0.575096
#> 920  0.918594 0.000010 0.745349
#> 921  0.999990 0.018038 0.999990
#> 922  0.999990 0.626344 0.999990
#> 923  0.897084 0.999990 0.588469
#> 924  0.999990 0.515023 0.999990
#> 925  0.999990 0.474612 0.999990
#> 926  0.999990 0.000010 0.999990
#> 927  0.000010 0.999990 0.646176
#> 928  0.999990 0.999990 0.705466
#> 929  0.999990 0.044949 0.999990
#> 930  0.744118 0.999990 0.682965
#> 931  0.999990 0.044949 0.999990
#> 932  0.999990 0.000010 0.720820
#> 933  0.999990 0.606095 0.999990
#> 934  0.954335 0.341087 0.796022
#> 935  0.999990 0.206421 0.510562
#> 936  0.999990 0.194277 0.558217
#> 937  0.999990 0.999990 0.692622
#> 938  0.999990 0.874981 0.999990
#> 939  0.999990 0.000010 0.999990
#> 940  0.999990 0.000010 0.999990
#> 941  0.999990 0.088894 0.668053
#> 942  0.999990 0.480269 0.764833
#> 943  0.999990 0.102728 0.999990
#> 944  0.999990 0.469109 0.999990
#> 945  0.999990 0.605699 0.999990
#> 946  0.868823 0.999990 0.366007
#> 947  0.418199 0.999990 0.999990
#> 948  0.936900 0.317699 0.999990
#> 949  0.999990 0.322946 0.449410
#> 950  0.059572 0.999990 0.827578
#> 951  0.521515 0.999990 0.999990
#> 952  0.000010 0.999990 0.796812
#> 953  0.999990 0.999990 0.667447
#> 954  0.029194 0.999990 0.697248
#> 955  0.969231 0.999990 0.651259
#> 956  0.970056 0.000010 0.488812
#> 957  0.999990 0.000010 0.999990
#> 958  0.999990 0.999990 0.626752
#> 959  0.999990 0.000010 0.929846
#> 960  0.999990 0.415539 0.968485
#> 961  0.081318 0.999990 0.999990
#> 962  0.999990 0.000010 0.952346
#> 963  0.999990 0.000010 0.952346
#> 964  0.999990 0.072391 0.558030
#> 965  0.999990 0.999990 0.679055
#> 966  0.999990 0.999990 0.305677
#> 967  0.999990 0.937654 0.721886
#> 968  0.000010 0.999990 0.999990
#> 969  0.000010 0.872242 0.859212
#> 970  0.000010 0.999990 0.658526
#> 971  0.000010 0.999990 0.652298
#> 972  0.999990 0.999990 0.574423
#> 973  0.999990 0.972125 0.568888
#> 974  0.999990 0.000010 0.999990
#> 975  0.109379 0.999990 0.999990
#> 976  0.902285 0.999990 0.714059
#> 977  0.999990 0.166116 0.238259
#> 978  0.555026 0.619365 0.396030
#> 979  0.788026 0.999990 0.393636
#> 980  0.213649 0.999990 0.557118
#> 981  0.519312 0.999990 0.013414
#> 982  0.999990 0.000010 0.999990
#> 983  0.999990 0.200047 0.999990
#> 984  0.999990 0.000010 0.999990
#> 985  0.967285 0.999990 0.663723
#> 986  0.967285 0.999990 0.663723
#> 987  0.966923 0.999990 0.678997
#> 988  0.000010 0.999990 0.698562
#> 989  0.999990 0.000010 0.999990
#> 990  0.999990 0.098198 0.999990
#> 991  0.999990 0.000010 0.749835
#> 992  0.999990 0.387454 0.999990
#> 993  0.999990 0.314892 0.999990
#> 994  0.000010 0.999990 0.963422
#> 995  0.999990 0.000010 0.999990
#> 996  0.999990 0.000010 0.999990
#> 997  0.999990 0.000010 0.999990
#> 998  0.999990 0.000010 0.999990
#> 999  0.999990 0.232707 0.999990
#> 1000 0.999990 0.182360 0.999990
#> 1001 0.999990 0.000010 0.604992
#> 1002 0.999990 0.000010 0.915171
#> 1003 0.999990 0.000010 0.915171
#> 1004 0.000010 0.999990 0.690879
#> 1005 0.999990 0.102728 0.999990
#> 1006 0.937315 0.999990 0.740304
#> 1007 0.999990 0.497739 0.080486
#> 1008 0.999990 0.586925 0.759349
#> 1009 0.999990 0.999990 0.518208
#> 1010 0.000010 0.999990 0.806271
#> 1011 0.999990 0.999990 0.668402
#> 1012 0.999990 0.000010 0.999990
#> 1013 0.999990 0.142738 0.999990
#> 1014 0.000010 0.999990 0.950304
#> 1015 0.999990 0.560540 0.999990
#> 1016 0.999990 0.000010 0.999990
#> 1017 0.189472 0.778452 0.745689
#> 1018 0.999990 0.000010 0.721627
#> 1019 0.999990 0.000010 0.999990
#> 1020 0.969136 0.000010 0.999990
#> 1021 0.999990 0.000010 0.999990
#> 1022 0.999990 0.000010 0.999990
#> 1023 0.999990 0.000010 0.999990
#> 1024 0.999990 0.000010 0.999990
#> 1025 0.234736 0.920942 0.963536
#> 1026 0.999990 0.081962 0.940874
#> 1027 0.999990 0.000010 0.999990
#> 1028 0.999990 0.000010 0.999990
#> 1029 0.999990 0.000010 0.999990
#> 1030 0.503834 0.345406 0.543156
#> 1031 0.999990 0.068642 0.999990
#> 1032 0.999990 0.999990 0.540337
#> 1033 0.186894 0.999990 0.799447
#> 1034 0.999990 0.713269 0.546763
#> 1035 0.999990 0.550362 0.999990
#> 1036 0.032128 0.806179 0.999990
#> 1037 0.999990 0.999990 0.304541
#> 1038 0.000010 0.953121 0.971204
#> 1039 0.999990 0.651014 0.968089
#> 1040 0.000010 0.999990 0.740014
#> 1041 0.999990 0.562745 0.999990
#> 1042 0.333118 0.999990 0.792432
#> 1043 0.999990 0.044949 0.999990
#> 1044 0.999990 0.343290 0.999990
#> 1045 0.999990 0.999990 0.559822
#> 1046 0.999990 0.999990 0.504972
#> 1047 0.999990 0.288723 0.999990
#> 1048 0.999990 0.000010 0.999990
#> 1049 0.999990 0.999990 0.643349
#> 1050 0.999990 0.000010 0.923089
#> 1051 0.999990 0.999990 0.606022
#> 1052 0.296001 0.795789 0.999990
#> 1053 0.999990 0.999990 0.621351
#> 1054 0.999990 0.000010 0.999990
#> 1055 0.999990 0.555504 0.999990
#> 1056 0.921770 0.671161 0.562632
#> 1057 0.072578 0.999990 0.667627
#> 1058 0.925711 0.999990 0.417148
#> 1059 0.999990 0.672043 0.683856
#> 1060 0.999990 0.388817 0.650540
#> 1061 0.548575 0.999990 0.849834
#> 1062 0.999990 0.951588 0.588686
#> 1063 0.933880 0.999990 0.757512
#> 1064 0.999990 0.999592 0.999990
#> 1065 0.999990 0.999990 0.437948
#> 1066 0.999990 0.999990 0.437948
#> 1067 0.000010 0.999990 0.870133
#> 1068 0.999990 0.999990 0.437948
#> 1069 0.999990 0.038438 0.999990
#> 1070 0.999990 0.038438 0.999990
#> 1071 0.999990 0.555505 0.999990
#> 1072 0.999990 0.000010 0.999990
#> 1073 0.999990 0.000010 0.999990
#> 1074 0.999990 0.999990 0.543116
#> 1075 0.999990 0.000010 0.999990
#> 1076 0.999990 0.501769 0.307524
#> 1077 0.000010 0.333332 0.999990
#> 1078 0.999990 0.498395 0.697072
#> 1079 0.999990 0.581510 0.875460
#> 1080 0.999990 0.944010 0.598332
#> 1081 0.059293 0.999990 0.999990
#> 1082 0.999990 0.200047 0.999990
#> 1083 0.000010 0.999990 0.845428
#> 1084 0.999990 0.049368 0.999990
#> 1085 0.999990 0.423597 0.999990
#> 1086 0.999990 0.000010 0.999990
#> 1087 0.999990 0.566573 0.999990
#> 1088 0.999990 0.000010 0.999990
#> 1089 0.999990 0.381679 0.942380
#> 1090 0.999990 0.381679 0.942380
#> 1091 0.999990 0.381679 0.942380
#> 1092 0.999990 0.369597 0.942417
#> 1093 0.967655 0.999990 0.382037
#> 1094 0.999990 0.462359 0.999990
#> 1095 0.999990 0.462359 0.999990
#> 1096 0.999990 0.191696 0.999990
#> 1097 0.000010 0.999990 0.959266
#> 1098 0.999990 0.351076 0.999990
#> 1099 0.999990 0.053140 0.904281
#> 1100 0.999990 0.999990 0.329024
#> 1101 0.000010 0.999990 0.917410
#> 1102 0.999990 0.000010 0.819662
#> 1103 0.999990 0.000010 0.999990
#> 1104 0.999990 0.999990 0.713006
#> 1105 0.999990 0.999990 0.624730
#> 1106 0.999990 0.000010 0.999990
#> 1107 0.272503 0.999990 0.999990
#> 1108 0.999990 0.000010 0.999990
#> 1109 0.999990 0.000010 0.381030
#> 1110 0.999990 0.446664 0.999990
#> 1111 0.999990 0.000010 0.772540
#> 1112 0.999990 0.999990 0.254004
#> 1113 0.999990 0.999990 0.365722
#> 1114 0.824158 0.999990 0.126732
#> 1115 0.999990 0.555950 0.999990
#> 1116 0.999990 0.000010 0.999990
#> 1117 0.999990 0.936699 0.399870
#> 1118 0.999990 0.000010 0.800519
#> 1119 0.409170 0.999990 0.999990
#> 1120 0.999990 0.190009 0.943978
#> 1121 0.999990 0.000010 0.999990
#> 1122 0.999990 0.000010 0.999990
#> 1123 0.999990 0.673776 0.736878
#> 1124 0.999990 0.362884 0.999990
#> 1125 0.999990 0.000010 0.999990
#> 1126 0.999990 0.000010 0.999990
#> 1127 0.999990 0.491335 0.869311
#> 1128 0.999990 0.000010 0.672471
#> 1129 0.000010 0.999990 0.742615
#> 1130 0.999990 0.362883 0.999990
#> 1131 0.964602 0.000010 0.690161
#> 1132 0.999990 0.362883 0.999990
#> 1133 0.999990 0.000010 0.887126
#> 1134 0.000010 0.999990 0.999990
#> 1135 0.052497 0.999990 0.913487
#> 1136 0.898062 0.999990 0.733584
#> 1137 0.999990 0.999990 0.434103
#> 1138 0.999990 0.776035 0.713206
#> 1139 0.999990 0.999990 0.697908
#> 1140 0.999990 0.343291 0.999990
#> 1141 0.999990 0.000010 0.886208
#> 1142 0.253511 0.999990 0.693575
#> 1143 0.999990 0.000010 0.999990
#> 1144 0.999990 0.000010 0.769674
#> 1145 0.999990 0.000010 0.769674
#> 1146 0.999990 0.000010 0.999990
#> 1147 0.999990 0.000010 0.450297
#> 1148 0.999990 0.000010 0.914436
#> 1149 0.999990 0.000010 0.999990
#> 1150 0.999990 0.917429 0.701237
#> 1151 0.999990 0.279974 0.999990
#> 1152 0.999990 0.158372 0.999990
#> 1153 0.304235 0.999990 0.999990
#> 1154 0.657161 0.999990 0.972624
#> 1155 0.895119 0.711118 0.297429
#> 1156 0.999990 0.279975 0.999990
#> 1157 0.999990 0.999990 0.694005
#> 1158 0.999990 0.000010 0.550598
#> 1159 0.999990 0.063626 0.999990
#> 1160 0.999990 0.898908 0.656558
#> 1161 0.999990 0.000010 0.539327
#> 1162 0.999990 0.625677 0.720879
#> 1163 0.744310 0.000010 0.999990
#> 1164 0.000010 0.999990 0.999990
#> 1165 0.000010 0.999990 0.999990
#> 1166 0.999990 0.372521 0.855218
#> 1167 0.980346 0.971158 0.595753
#> 1168 0.320715 0.999990 0.999990
#> 1169 0.999990 0.999990 0.535923
#> 1170 0.999990 0.471255 0.930538
#> 1171 0.355023 0.999990 0.848177
#> 1172 0.000010 0.999990 0.643740
#> 1173 0.000010 0.999990 0.999990
#> 1174 0.999990 0.000010 0.999990
#> 1175 0.999990 0.999990 0.669343
#> 1176 0.038427 0.999990 0.737133
#> 1177 0.235372 0.999990 0.999990
#> 1178 0.999990 0.000010 0.999990
#> 1179 0.000010 0.999990 0.677853
#> 1180 0.999990 0.680858 0.162306
#> 1181 0.283209 0.961167 0.696232
#> 1182 0.273721 0.999990 0.999990
#> 1183 0.000010 0.999990 0.999990
#> 1184 0.999990 0.000010 0.563539
#> 1185 0.999990 0.000010 0.513429
#> 1186 0.999990 0.742722 0.563843
#> 1187 0.955959 0.793900 0.895518
#> 1188 0.000010 0.999990 0.968091
#> 1189 0.999990 0.000010 0.970088
#> 1190 0.999990 0.108746 0.999990
#> 1191 0.999990 0.083315 0.999990
#> 1192 0.000010 0.999990 0.999990
#> 1193 0.999990 0.000010 0.790100
#> 1194 0.999990 0.254922 0.999990
#> 1195 0.999990 0.999990 0.643349
#> 1196 0.999990 0.254922 0.999990
#> 1197 0.999990 0.092089 0.999990
#> 1198 0.999990 0.092089 0.999990
#> 1199 0.858426 0.999990 0.831002
#> 1200 0.999990 0.299474 0.999990
#> 1201 0.999990 0.999990 0.468084
#> 1202 0.037590 0.999990 0.999990
#> 1203 0.999990 0.967248 0.705962
#> 1204 0.442679 0.799318 0.562166
#> 1205 0.556978 0.836565 0.442669
#> 1206 0.548288 0.830855 0.591913
#> 1207 0.999990 0.241744 0.999990
#> 1208 0.999990 0.999990 0.700574
#> 1209 0.999990 0.000010 0.999990
#> 1210 0.999990 0.000010 0.999990
#> 1211 0.000010 0.999990 0.833520
#> 1212 0.999990 0.266005 0.730149
#> 1213 0.999990 0.999990 0.681956
#> 1214 0.999990 0.000010 0.999990
#> 1215 0.999990 0.214762 0.999990
#> 1216 0.999990 0.083315 0.999990
#> 1217 0.999990 0.999990 0.535545
#> 1218 0.106961 0.471817 0.986257
#> 1219 0.999990 0.724809 0.398308
#> 1220 0.000010 0.999990 0.999990
#> 1221 0.999990 0.772412 0.999990
#> 1222 0.968850 0.496540 0.716293
#> 1223 0.999990 0.473801 0.999990
#> 1224 0.999990 0.743369 0.553455
#> 1225 0.999990 0.999990 0.609006
#> 1226 0.999990 0.000010 0.999990
#> 1227 0.999990 0.256659 0.999990
#> 1228 0.999990 0.044944 0.999990
#> 1229 0.999990 0.999990 0.696510
#> 1230 0.000010 0.957147 0.999990
#> 1231 0.999990 0.999990 0.620433
#> 1232 0.999990 0.000010 0.297868
#> 1233 0.999990 0.999990 0.544002
#> 1234 0.999990 0.999990 0.508234
#> 1235 0.999990 0.138509 0.999990
#> 1236 0.043306 0.523858 0.471590
#> 1237 0.061250 0.589832 0.330508
#> 1238 0.062063 0.486865 0.550906
#> 1239 0.999990 0.597344 0.999990
#> 1240 0.999990 0.472269 0.970193
#> 1241 0.377829 0.999990 0.274740
#> 1242 0.000048 0.010887 0.999990
#> 1243 0.999990 0.568936 0.999990
#> 1244 0.999990 0.000010 0.837994
#> 1245 0.999990 0.999990 0.615240
#> 1246 0.999990 0.000010 0.968934
#> 1247 0.999990 0.102618 0.501573
#> 1248 0.999990 0.368337 0.999990
#> 1249 0.999990 0.563217 0.999990
#> 1250 0.744605 0.999990 0.899212
#> 1251 0.999990 0.375848 0.999990
#> 1252 0.211734 0.999990 0.972833
#> 1253 0.999990 0.000010 0.757460
#> 1254 0.668753 0.874973 0.331247
#> 1255 0.347410 0.957083 0.999990
#> 1256 0.274403 0.999990 0.903844
#> 1257 0.999990 0.999990 0.390546
#> 1258 0.999990 0.098198 0.999990
#> 1259 0.937312 0.999990 0.681682
#> 1260 0.999990 0.000010 0.999990
#> 1261 0.000010 0.999990 0.699323
#> 1262 0.999990 0.422741 0.999990
#> 1263 0.999990 0.999990 0.604655
#> 1264 0.927911 0.677434 0.891524
#> 1265 0.031663 0.999990 0.999990
#> 1266 0.999990 0.454313 0.777453
#> 1267 0.176424 0.999990 0.999990
#> 1268 0.999990 0.038546 0.393077
#> 1269 0.999990 0.038546 0.393077
#> 1270 0.000010 0.999990 0.999990
#> 1271 0.999990 0.080387 0.950665
#> 1272 0.999990 0.013441 0.999990
#> 1273 0.000010 0.999990 0.682644
#> 1274 0.999990 0.999990 0.252673
#> 1275 0.999990 0.999990 0.290774
#> 1276 0.999990 0.000010 0.456575
#> 1277 0.896008 0.185322 0.838060
#> 1278 0.999990 0.000010 0.999990
#> 1279 0.999990 0.343291 0.999990
#> 1280 0.999990 0.343291 0.999990
#> 1281 0.999990 0.343291 0.999990
#> 1282 0.999990 0.343291 0.999990
#> 1283 0.999990 0.000010 0.555221
#> 1284 0.999990 0.999990 0.540352
#> 1285 0.999990 0.917359 0.193774
#> 1286 0.881813 0.999990 0.563428
#> 1287 0.881813 0.999990 0.563428
#> 1288 0.999990 0.999990 0.687228
#> 1289 0.999990 0.000010 0.999990
#> 1290 0.999990 0.516735 0.999990
#> 1291 0.999990 0.187829 0.768728
#> 1292 0.999990 0.000010 0.999990
#> 1293 0.000010 0.999990 0.999990
#> 1294 0.999990 0.059597 0.999990
#> 1295 0.999990 0.000010 0.948116
#> 1296 0.999990 0.999990 0.648485
#> 1297 0.999990 0.349284 0.999990
#> 1298 0.103298 0.859317 0.981351
#> 1299 0.999990 0.560540 0.999990
#> 1300 0.290612 0.999990 0.491068
#> 1301 0.999990 0.168434 0.328417
#> 1302 0.000010 0.999990 0.910385
#> 1303 0.999990 0.000010 0.926162
#> 1304 0.999990 0.142736 0.999990
#> 1305 0.695964 0.999990 0.937732
#> 1306 0.911940 0.475459 0.835790
#> 1307 0.402814 0.999990 0.357828
#> 1308 0.000010 0.999990 0.999990
#> 1309 0.999990 0.114343 0.695263
#> 1310 0.999990 0.259254 0.971564
#> 1311 0.037558 0.999990 0.813605
#> 1312 0.037747 0.999990 0.815011
#> 1313 0.953044 0.999990 0.431464
#> 1314 0.041983 0.999990 0.706837
#> 1315 0.746401 0.999990 0.837892
#> 1316 0.999990 0.355312 0.845774
#> 1317 0.000010 0.999990 0.803831
#> 1318 0.930287 0.999990 0.775461
#> 1319 0.000010 0.999990 0.648905
#> 1320 0.999990 0.000010 0.999990
#> 1321 0.000010 0.999990 0.896605
#> 1322 0.999990 0.999990 0.198099
#> 1323 0.999990 0.563217 0.999990
#> 1324 0.999990 0.148317 0.857055
#> 1325 0.999990 0.000010 0.744744
#> 1326 0.999990 0.148397 0.851648
#> 1327 0.000010 0.999990 0.708459
#> 1328 0.999990 0.515604 0.999990
#> 1329 0.999990 0.515604 0.999990
#> 1330 0.999990 0.000010 0.972057
#> 1331 0.132474 0.999990 0.999990
#> 1332 0.999990 0.000010 0.397480
#> 1333 0.999990 0.607526 0.999990
#> 1334 0.999990 0.999990 0.245096
#> 1335 0.000010 0.999990 0.943817
#> 1336 0.999990 0.999990 0.713078
#> 1337 0.077031 0.999990 0.999990
#> 1338 0.999990 0.098448 0.968820
#> 1339 0.043708 0.999990 0.835625
#> 1340 0.999990 0.000010 0.999990
#> 1341 0.999990 0.469109 0.999990
#> 1342 0.999990 0.000010 0.675038
#> 1343 0.999990 0.012029 0.834094
#> 1344 0.999990 0.309632 0.912132
#> 1345 0.999990 0.999990 0.580427
#> 1346 0.937366 0.895855 0.313153
#> 1347 0.060478 0.999990 0.725169
#> 1348 0.999990 0.999990 0.138744
#> 1349 0.000010 0.999990 0.864796
#> 1350 0.999990 0.905628 0.746301
#> 1351 0.999990 0.000010 0.870463
#> 1352 0.000010 0.999990 0.999990
#> 1353 0.999990 0.999990 0.254416
#> 1354 0.999990 0.049143 0.337218
#> 1355 0.042887 0.999990 0.664720
#> 1356 0.490834 0.728926 0.511738
#> 1357 0.999990 0.999990 0.642388
#> 1358 0.999990 0.422742 0.999990
#> 1359 0.999990 0.219821 0.936007
#> 1360 0.999990 0.000010 0.999990
#> 1361 0.000065 0.991682 0.995603
#> 1362 0.999990 0.000010 0.999990
#> 1363 0.999990 0.468148 0.999990
#> 1364 0.999990 0.381540 0.999990
#> 1365 0.999990 0.000010 0.373702
#> 1366 0.999990 0.000010 0.944020
#> 1367 0.999990 0.560540 0.999990
#> 1368 0.999990 0.000010 0.944020
#> 1369 0.999990 0.318185 0.810403
#> 1370 0.999990 0.999990 0.582177
#> 1371 0.999990 0.999990 0.616049
#> 1372 0.133633 0.999990 0.742475
#> 1373 0.999990 0.074703 0.999990
#> 1374 0.999990 0.000010 0.999990
#> 1375 0.999990 0.000010 0.515041
#> 1376 0.999990 0.617479 0.528974
#> 1377 0.999990 0.340572 0.727946
#> 1378 0.030264 0.948888 0.832640
#> 1379 0.517093 0.271305 0.733470
#> 1380 0.934618 0.249087 0.426336
#> 1381 0.999990 0.941495 0.765543
#> 1382 0.999990 0.956711 0.757594
#> 1383 0.000010 0.999990 0.798453
#> 1384 0.999990 0.209235 0.879479
#> 1385 0.999990 0.492689 0.767097
#> 1386 0.999990 0.000010 0.702154
#> 1387 0.999990 0.999990 0.645280
#> 1388 0.000010 0.999990 0.944915
#> 1389 0.999990 0.000010 0.576138
#> 1390 0.999990 0.000010 0.714593
#> 1391 0.999990 0.000010 0.697957
#> 1392 0.999990 0.499995 0.999990
#> 1393 0.999990 0.000010 0.760050
#> 1394 0.999990 0.295626 0.878991
#> 1395 0.999990 0.000010 0.799770
#> 1396 0.999990 0.402703 0.950779
#> 1397 0.999990 0.405465 0.482652
#> 1398 0.999990 0.519465 0.913433
#> 1399 0.999990 0.077395 0.375384
#> 1400 0.999990 0.429324 0.999990
#> 1401 0.999990 0.555954 0.999990
#> 1402 0.999990 0.999990 0.539651
#> 1403 0.999990 0.515604 0.999990
#> 1404 0.999990 0.515604 0.999990
#> 1405 0.999990 0.000010 0.999990
#> 1406 0.999990 0.605698 0.999990
#> 1407 0.999990 0.999990 0.710554
#> 1408 0.999990 0.098198 0.999990
#> 1409 0.409620 0.999990 0.999990
#> 1410 0.999990 0.473501 0.908462
#> 1411 0.999990 0.174711 0.999990
#> 1412 0.999990 0.000010 0.999990
#> 1413 0.999990 0.047996 0.999990
#> 1414 0.999990 0.355996 0.939649
#> 1415 0.999990 0.061815 0.940584
#> 1416 0.999990 0.061815 0.940584
#> 1417 0.999990 0.999990 0.547360
#> 1418 0.999990 0.331231 0.999990
#> 1419 0.671041 0.960509 0.692016
#> 1420 0.999990 0.043935 0.815723
#> 1421 0.999990 0.000010 0.814417
#> 1422 0.999990 0.343849 0.999990
#> 1423 0.999990 0.999990 0.499984
#> 1424 0.999990 0.000010 0.999990
#> 1425 0.999990 0.470289 0.999990
#> 1426 0.999990 0.276546 0.455315
#> 1427 0.999990 0.470289 0.999990
#> 1428 0.999990 0.999990 0.552605
#> 1429 0.999990 0.049631 0.943305
#> 1430 0.999990 0.560540 0.999990
#> 1431 0.999990 0.311643 0.999990
#> 1432 0.999990 0.266869 0.999990
#> 1433 0.999990 0.000010 0.690884
#> 1434 0.999990 0.999990 0.714150
#> 1435 0.999990 0.142738 0.999990
#> 1436 0.999990 0.287169 0.999990
#> 1437 0.999990 0.049821 0.906238
#> 1438 0.999990 0.049659 0.937583
#> 1439 0.000010 0.999990 0.856660
#> 1440 0.999990 0.999990 0.667450
#> 1441 0.999990 0.999990 0.682280
#> 1442 0.000010 0.999990 0.821699
#> 1443 0.999990 0.913192 0.766401
#> 1444 0.999990 0.913192 0.766401
#> 1445 0.337015 0.999990 0.999990
#> 1446 0.337015 0.999990 0.999990
#> 1447 0.999990 0.000010 0.662482
#> 1448 0.999990 0.000010 0.835684
#> 1449 0.999990 0.596967 0.999990
#> 1450 0.999990 0.000010 0.666017
#> 1451 0.999990 0.803922 0.705266
#> 1452 0.999990 0.000010 0.710633
#> 1453 0.999990 0.999990 0.639100
#> 1454 0.999990 0.555505 0.999990
#> 1455 0.176642 0.999990 0.955360
#> 1456 0.999990 0.000010 0.874376
#> 1457 0.999990 0.070274 0.902890
#> 1458 0.884420 0.999990 0.715655
#> 1459 0.999990 0.000010 0.999990
#> 1460 0.000010 0.999990 0.999990
#> 1461 0.999990 0.000010 0.571852
#> 1462 0.999990 0.000010 0.571852
#> 1463 0.000010 0.999990 0.999990
#> 1464 0.999990 0.575979 0.908851
#> 1465 0.999990 0.886726 0.379380
#> 1466 0.999990 0.555504 0.999990
#> 1467 0.999990 0.000010 0.444055
#> 1468 0.999990 0.433501 0.999990
#> 1469 0.832500 0.000010 0.928777
#> 1470 0.847067 0.000010 0.838461
#> 1471 0.999990 0.000010 0.999990
#> 1472 0.999990 0.999990 0.694005
#> 1473 0.999990 0.000010 0.999990
#> 1474 0.999990 0.000010 0.759364
#> 1475 0.999990 0.106740 0.772772
#> 1476 0.999990 0.000010 0.568734
#> 1477 0.310989 0.999990 0.999990
#> 1478 0.000010 0.999990 0.999990
#> 1479 0.999990 0.000010 0.723999
#> 1480 0.999990 0.000010 0.723999
#> 1481 0.390505 0.703302 0.921023
#> 1482 0.999990 0.000010 0.898081
#> 1483 0.000010 0.999990 0.999990
#> 1484 0.999990 0.000010 0.783844
#> 1485 0.029955 0.999990 0.642602
#> 1486 0.999990 0.999990 0.124633
#> 1487 0.999990 0.929549 0.216924
#> 1488 0.794617 0.999990 0.788390
#> 1489 0.999990 0.000010 0.999990
#> 1490 0.999990 0.000010 0.999990
#> 1491 0.452801 0.999990 0.999990
#> 1492 0.999990 0.759615 0.621951
#> 1493 0.999990 0.098198 0.999990
#> 1494 0.999990 0.999990 0.565971
#> 1495 0.999990 0.000010 0.770176
#> 1496 0.999990 0.279974 0.999990
#> 1497 0.913595 0.000010 0.609275
#> 1498 0.999990 0.000010 0.999990
#> 1499 0.000010 0.999990 0.866466
#> 1500 0.999990 0.999990 0.648217