Save a gen_tibble

Save the tibble (and update the backing files). The gen_tibble object is saved to a file with extension .gt, togethe with update its .rds and .bk files. Note that multiple .gt files can be linked to the same .rds and .bk files; generally, this occurs when we create multiple subsets of the data. The .gt file then stores the information on what subset of the full dataset we are interested in, whilst the .rds and .bk file store the full dataset. To reload a gen_tibble, you can pass the name of the .gt file with gt_load().

Usage

gt_save(x, file_name = NULL, quiet = FALSE)

Arguments

x

a gen_tibble

file_name

the file name, including the full path. If it does not end with .gt, the extension will be added.

quiet

boolean to suppress information about the files

Value

the file name and path of the .gt file, together with the .rds and .bk files

See also