Estimate allele frequencies at each each locus

Allele frequencies can be estimates as minimum allele frequencies (MAF) with loci_maf() or the frequency of the alternate allele (with loci_alt_freq()). The latter are in line with the genotypes matrix (e.g. as extracted by show_loci()). Most users will be in interested in the MAF, but the raw frequencies might be useful when computing aggregated statistics.

loci_alt_freq(.x, ...)

# S3 method for tbl_df
loci_alt_freq(.x, ...)

# S3 method for vctrs_bigSNP
loci_alt_freq(.x, ...)

# S3 method for grouped_df
loci_alt_freq(.x, n_cores = bigstatsr::nb_cores(), ...)

loci_maf(.x, ...)

# S3 method for tbl_df
loci_maf(.x, ...)

# S3 method for vctrs_bigSNP
loci_maf(.x, ...)

# S3 method for grouped_df
loci_maf(.x, n_cores = bigstatsr::nb_cores(), ...)

Arguments

.x

a vector of class vctrs_bigSNP (usually the genotypes column of a gen_tibble object), or a gen_tibble.

...

other arguments passed to specific methods, currently unused.

n_cores

number of cores to be used, it defaults to bigstatsr::nb_cores()

Value

a vector of frequencies, one per locus