Compute the square of the Frobenius norm of a matrix

This function computes the square of the Frobenius norm of a matrix, which is the sum of the squares of the matrix elements, which provides a measure of of the total variance of the matrix. The code used here was outlined in an issue of bigstatsr by @privefl: https://github.com/privefl/bigstatsr/issues/83

Usage

square_frobenious(
  X,
  ind.row = bigstatsr::rows_along(X),
  ind.col = bigstatsr::cols_along(X),
  center = rep(0, length(ind.col)),
  scale = rep(1, length(ind.col))
)

Arguments

X

An FBM matrix

ind.row

A vector of row indices

ind.col

A vector of column indices

center

A vector of centering values (i.e the means of the genotype counts)

scale

A vector of scaling values (i.e the standard deviations of the genotype counts)

Value

The square of the Frobenius norm of the matrix