Tidy a Q matrix

Takes a q_matrix object, which is a matrix, and returns a tidied tibble.

Usage

# S3 method for class 'q_matrix'
tidy(x, data, ...)

Arguments

x

A Q matrix object (as returned by q_matrix).

data

An associated tibble (e.g. a gen_tibble), with the individuals in the same order as the data used to generate the Q matrix

...

not currently used

Value

A tidied tibble containing columns:

row

ID of the original observation (i.e. rowname from original data).

Q

Integer indicating a Q component.

value

The proportion for that particular Q value.