Convert vcf to FBM.

Convert a vcf file to a Filebacked Big Matrix (FBM) object. This should work even for large vcf files that would not fit in memory.

Usage

vcf_to_fbm_cpp(vcf_path, backingfile = NULL, quiet = FALSE)

Arguments

vcf_path

the path to the vcf

backingfile

the name of the file to use as the backing file for the FBM. If NULL, the vcf path will be used.

Value

path to the resulting rds file as class bigSNP.