Convert vcf to FBM.

Convert a vcf file to a Filebacked Big Matrix (FBM) object. This should work even for large vcf files that would not fit in memory.

Usage

vcf_to_fbm_cpp(
  vcf_path,
  chunk_size = NULL,
  backingfile = NULL,
  n_cores = 1,
  quiet = FALSE
)

Arguments

vcf_path

the path to the vcf

chunk_size

the chunk size to use on the vcf when loading the file. If NULL, a best guess will be made.

backingfile

the name of the file to use as the backing file for the FBM. If NULL, the vcf path will be used.

n_cores

the number of cores to use when reading the vcf file. Default is 1.

Value

path to the resulting rds file as class bigSNP.